Moved student_list.dart
+ fixed list button on addStudentPage. List button doesn't need to add students
